requirements.txt
langchain==1.2.10
langchain-core==1.2.12
langchain-community==0.4.1
langchain-openai==1.1.9
langchain-text-splitters==1.1.0
openai==2.6.1
chromadb==1.4.0
pypdf==6.5.0
python-dotenv==1.2.1
tiktoken==0.12.0